remote-hg: always try to initialize shared repo
We don't need the initial clone, if the repository is shared, pulling
from the child updates the parent's storage; it's exactly the same as
cloning.

This way the code is simpler.

+++ contrib/remote-helpers/git-remote-hg.py
@@ -391,24 +391,22 @@ def get_repo(url, alias):
             os.makedirs(dirname)
     else:
         shared_path = os.path.join(gitdir, 'hg')
-        if not os.path.exists(shared_path):
-            try:
-                hg.clone(myui, {}, url, shared_path, update=False, pull=True)
-            except:
-                die('Repository error')
-        else:
-            # check and upgrade old organization
-            hg_path = os.path.join(shared_path, '.hg')
-            if not os.path.exists(hg_path):
-                repos = os.listdir(shared_path)
-                for x in repos:
-                    local_hg = os.path.join(shared_path, x, 'clone', '.hg')
-                    if not os.path.exists(local_hg):
-                        continue
-                    shutil.copytree(local_hg, hg_path)
-            # no clone found, just init
-            if not os.path.exists(hg_path):
-                hg.peer(myui, {}, shared_path, create=True)
+
+        # check and upgrade old organization
+        hg_path = os.path.join(shared_path, '.hg')
+        if os.path.exists(shared_path) and not os.path.exists(hg_path):
+            repos = os.listdir(shared_path)
+            for x in repos:
+                local_hg = os.path.join(shared_path, x, 'clone', '.hg')
+                if not os.path.exists(local_hg):
+                    continue
+                shutil.copytree(local_hg, hg_path)
+
+        # setup shared repo (if not there)
+        try:
+            hg.peer(myui, {}, shared_path, create=True)
+        except error.RepoError:
+            pass
 
         if not os.path.exists(dirname):
             os.makedirs(dirname)
